fix compilation error (not real fix): error C4146: unary minus operator applied to unsigned type, result still unsigned

+++ src/tccasm.c
@@ -86,7 +86,7 @@ static void asm_expr_unary(TCCState *s1, ExprValue *pe)
         if (pe->sym)
             tcc_error("invalid operation with label");
         if (op == '-')
-            pe->v = -pe->v;
+            pe->v = -(int)pe->v;
         else
             pe->v = ~pe->v;
         break;
